requirements.txt
apifairy==0.9.2
apispec==5.2.2
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-HTTPAuth==4.7.0
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.39
webargs==8.2.0
Werkzeug==2.1.2
